[Refactor] Move user task startup logic from kernel/usermain.c to usermain.c
- Move all user-defined task creation and startup code to standalone usermain.c
- Keep kernel/usermain.c minimal with only a weak definition of usermain()
- Add usermain.c to executable sources via CMake for better modularity
- Improves separation of user task logic from kernel internals
